+++ spec/better_array_spec.js
@@ -1,17 +1,17 @@
 'use strict';
 
-if (typeof(betterArray) === 'undefined') {
+if (typeof(BetterArray) === 'undefined') {
 	  var isNode = true;
-    var a = require("../better-array");
+    var ba = require("../better-array");
 } else {
 	  var isNode = false;
-		var a = betterArray;
+		var ba = BetterArray;
 }
 
 describe('better-array', function() {
 	describe('#minus', function() {
     it("is like Ruby's Array#-", function() {
-      expect( a([2,3,4]).minus([3,4,5]) ).toEqual([2]);
+      expect( ba([2,3,4]).minus([3,4,5]) ).toEqual([2]);
     });
 	});
 });