+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v0.2.0
 
 * New chain method
+* Rename to BetterArray
 
 ## v0.1.0